+++ examples/react-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # editful-react-plugin-example
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @editful/canvas-sdk@0.9.2
+
 ## 1.0.1
 
 ### Patch Changes