+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @editful/cli
 
+## 0.9.2
+
 ## 0.9.1
 
 ### Patch Changes