+++ packages/plugin-artifact/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @editful/plugin-artifact
 
+## 0.9.2
+
 ## 0.9.1
 
 ## 0.9.0